+++ CHANGELOG.md
@@ -22,6 +22,8 @@ More information can be found in our [Performance](https://docs.sentry.io/produc
 - [BC BREAK] Remove the `Context::CONTEXT_USER`, `Context::CONTEXT_RUNTIME`, `Context::CONTEXT_TAGS`, `Context::CONTEXT_EXTRA`, `Context::CONTEXT_SERVER_OS` constants (#1047)
 - [BC BREAK] Use PSR-17 factories in place of the Httplug's ones and return a promise from the transport (#1066)
 - [BC BREAK] The Monolog handler does not set anymore tags and extras on the event object (#1068)
+- [BC BREAK] Remove the `UserContext`, `ExtraContext` and `Context` classes and refactor the `ServerOsContext` and `RuntimeContext` classes (#1071)
+- Add the `EnvironmentIntegration` integration to gather data for the `os` and `runtime` contexts (#1071)
 
 ### 2.4.3 (2020-08-13)